fix clear fichier

+++ src/pages/DetailsRecipes.jsx
@@ -1,15 +1,17 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Link, useHistory } from 'react-router-dom'
 import { Flex, Box, chakra, Image, Button, Stack, List, ListItem, ListIcon} from '@chakra-ui/react';
-import {AiFillDelete, AiFillEdit} from 'react-icons/ai'
+import {AiFillEdit} from 'react-icons/ai'
 import {GiCookingPot, GiCheckMark} from 'react-icons/gi'
 import{ useParams } from 'react-router-dom';
 import Loader from '../utils/Loader';
 import axios from 'axios';
 import { useFetch } from '../utils/hooks/useFetch';
+import ConfirmDelete from '../components/ConfirmDelete';
 
 const DetailsRecipes = () => {
     const history = useHistory()
+    const [btnSup, setBtnSup] = useState(false)
     const params = useParams();
     const id = params.id;
     const { data, error, isLoading } = useFetch(`http://localhost:9000/api/recipe/${id}`);    
@@ -18,15 +20,15 @@ const DetailsRecipes = () => {
     if (error) {
       return  <span>{error}</span>
     }
-    const onDeleteRecipe = (id) => {
-        axios.delete(`http://localhost:9000/api/recipe/${id}`)
-        .then(resul => {
-            alert(resul.data.message)
-            history.push('/')
-        })
-        .catch((e) => {
-           alert(e); 
-        })
+    const onDeleteRecipe = (btnSup, id) => {
+        if(btnSup === true) {
+            axios.delete(`http://localhost:9000/api/recipe/${id}`)
+            .then(resul => {
+                // alert(resul.data.message)
+                setBtnSup(false)
+                history.push('/')
+            })   
+        }   
     }
     return (
             <>
@@ -35,10 +37,9 @@ const DetailsRecipes = () => {
             ) : ( recipe &&
                 <Flex
                     p={20}
-                    key={recipe.id}
                     w="full"
                     alignItems="center"
-                    justifyContent="center"    
+                    justifyContent="center"  
                 >
                     <Box
                         mx="auto"
@@ -46,8 +47,8 @@ const DetailsRecipes = () => {
                         shadow="md"
                         maxW="2xl"
                         mt={14}
+                        key={recipe.id}
                     >
-
                         <Image
                         roundedTop="lg"
                         w="full"
@@ -56,19 +57,17 @@ const DetailsRecipes = () => {
                         src={recipe?.photo}
                         alt={recipe?.titre}
                         />
-
                     <Box p={6}>
                 <Box>
                     <chakra.h2
-                    fontSize={{ base: "2xl", md: "3xl" }}
-                    textTransform="uppercase"
+                        fontSize={{ base: "2xl", md: "3xl" }}
+                        textTransform="uppercase"
                     >
                     {recipe.titre}
                     </chakra.h2>
-                    
                     <chakra.p
-                    mt={2}
-                    fontSize="sm"
+                        mt={2}
+                        fontSize="sm"
                     >
                     {recipe?.description}
                     </chakra.p>
@@ -83,11 +82,9 @@ const DetailsRecipes = () => {
                             {recipe?.personnes}{' '} personne{recipe?.personnes > 0 && 's'}
                         </chakra.span>
                     </Stack>
-                    
                     <chakra.h3 my={2}>
                         INGREDIENTS :
                     </chakra.h3>
-        
                     <List spacing={3}>
                         {recipe.ingredients?.map(ingredient => 
                         <ListItem>
@@ -112,27 +109,14 @@ const DetailsRecipes = () => {
                         <Button leftIcon={<AiFillEdit />} colorScheme="blue" variant="outline">
                         <Link to={`../edit/${recipe.id}`}>Modifier</Link>
                         </Button>
-                        <Button 
-                            onClick={() =>{
-                                const confirBox = window.confirm(
-                                    `voulez-vous supprimer cette recette ${recipe.titre}`
-                                )
-                                if (confirBox === true) {
-                                    onDeleteRecipe(recipe.id)
-                                }  
-                            }}
-                            rightIcon={<AiFillDelete />} 
-                            colorScheme="red" variant="outline"
-                        >
-                            Supprimer
-                        </Button>
+                        <ConfirmDelete onDeleteRecipe={onDeleteRecipe} onId={recipe.id} />
                     </Stack>
                 </Box>
                 </Box>
             </Box>
-            </Flex>  
-            )}
-        </>
+        </Flex>  
+        )}
+    </>
     )
 }
